Refactoring specs. TESTS STILL BROKEN
