+++ i18n_country_select.gemspec
@@ -13,7 +13,9 @@ Gem::Specification.new do |s|
   s.require_path = 'lib'
   s.rubyforge_project = '[none]'
   s.required_rubygems_version = '>= 1.3.5'
-  s.add_dependency('i18n-country-translations', ">= 0.0.1")
+  s.add_dependency('i18n', '~> 0.5')
+  s.add_dependency('i18n-country-translations', ">= 0.0.2")
   s.add_development_dependency "rails", ">= 3.0.0"
   s.add_development_dependency "rspec-rails", ">= 2.7.0"
+  s.add_development_dependency "spork", "~> 1.0rc"
 end
