+++ lib/i18n_country_select.rb
@@ -1,3 +1,4 @@
+require "i18n-country-translations"
 require "i18n_country_select/countries"
 require "i18n_country_select/form_builder"
 require "i18n_country_select/form_helpers"
